some component arrangement change
