+++ app/(meta)/privacy/page.tsx
@@ -1,0 +1,3 @@
+export default function Privacy() {
+    return <div></div>;
+}
