--- app/_components/spotlight.tsx
@@ -1,3 +1,0 @@
-export default function Spotlight() {
-  return <div className="px-4 lg:px-8 py-6"></div>;
-}
